Improve RollingUpdate detection
issues: #4

In order to detect a PodTemplateSpec update in the `RedisCluster`
custom resource instance, the operator generates an md5 hash from
the `RedisCluster.spec.template.spec` and add this hash in an
annotation for each RedisNode pod create from this `PodSpec`.

Thanks to that if a PodTemplateSpec is updated later in the
`RedisCluster` custom resource. The Operator will be able to detect
that a rolling-update is needed.