New --config-file option in redis-node (#22)
Fixes: #19

Add new redis-node option for including redis conf file in the generated
redis-conf file.
The file will be added at the end of the default file. like this those
additional configuration files will be able to overwrite default values
or value already set.

Redis-Cluster helm chart has been updated in order to benefit from the
new parameter. Now when a user wants to provide additional redis-server
configuration, it can be done with the parameters stored in a file in a
ConfigMap. that config file needs to be provided to the "helm install"
command, by setting your file path in the value:
"redis.configuration.file". Be carefull the file must be stored in the
chart path and the file path is relative to the chart root path.

Also in the chart now the redis-server configuration is generated in an
emptyDir in order to support "read only root file system" container
configuration like it can be the case in some kubernetes or openshift
cluster configuration.